refactor utils package
